requirements.txt
annotated-types==0.7.0
audioop-lts==0.2.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docopt==0.6.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
pipreqs==0.4.13
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
requests==2.32.3
rsa==4.9
sh==2.2.1
SpeechRecognition==3.14.1
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
yarg==0.1.10
